Add message & /command signals to edge calculation
Uses the shiny new logging system and data model to compute directed edge weights between users.

Uses mostly-arbitrary weighting functions to create a total relationship weight from messages sent as well as topic changes, whispers, invites, and kicks.

Also, get rid of the Sender and Recipient objects that were just wrappers around defaultdicts - it's easier/cleaner to use the straight dictionaries, and it's also easier to print for debugging.

Next I'm going to use the edge weights to update the current vinebots with /new and /del_friendship.

Note that there's some hacky stuff going on to filter out failed commands. Since the commands db entry is already inserted by the time we actually try to execute the command, it has no way to know if it's failed, and the best it can do is pretend to have succeeded so that handle_msg doesn't create a *second* entry for that same command, one with is_valid=true and one with is_valid=false. Later I might update the stored command's is_successful column (which doesn't exist yet), but the extra database write call seems like it's not worth the complexity for now. In the meantime, just remember that messages in response to /commands that start with "Sorry, " might be excluded from edge weight calculations!