Finish first version of edge scoring script
calc_edges.py is intended to be run once a day at an off-peak time. It looks at all of the activity on Vine (and, eventually, other services) over the past 90 days, and uses that to determine the relative weigths of directed relationships between each pair of users. It then looks at the edges that were in the database at the time of the start of the script, and updates them to reflect the current relationships, and it uses the /new and /del_friendship commands to update the leaf vinebots accordingly.

Note that there is currently some weird/buggy behavior that I *think* results from the fact that the script is designed for asymmetric relationships, and the leaf is designed for symmetric relationships. The next big task is to refactor all of the vinebot code to support asymmetric relationships. Hopefully, while I'm rewriting everything, I can also update it to use database transactions so that multiple leaves can run simultaneously.