Add instructions and an installation shell script
Installing JWChat from the repository source requires multiple steps, so it seemed easiest to encapsulate it in a little script. First, this script configures the server, username, and password. Then, it optionally checks out the Debugger files, and enables the flag. Finall,y, it removes any previous files from the nginx directory, and transfers over the new ones.