Temporarily serve a dynamic jwchat config file
But I don't think I'll keep it this way. If the config needs the password to log in (which I'm not sure about yet), then it can't be served on a separate request from the demo page itself, that has the secret URL. If the config file only needs the server name, then there's an easier way to do serve it, perhaps with separate files for dev and prod, both kept in the vine branch of JWChat.