Re-add Twitter auth from master for existing users
This is largely just a copy-paste of snippets from the master branch. Signing in only works for Twitter users who have a username that matches one in the database.

Also, this fixes the Gunicorn Twitter auth issue -Gunicorn was complaining because the secret key wasn't set, and it was complaining silently because the debug flag wasn't set to true, and this was because the __main__ code was never getting run when the app was started with Gunicorn, but it was getting run when the app was started with Flask's built-in webserver.