Fix everything, update shared repos with constants
Ugh, so I wanted to push a quick nginx change to prod, because I was noticing 500 errors for what *looked* to be the first four times the page was loaded, which matches the number of gunicorn workers I had, so I thought that maybe it was timing out if the database connectionns were dropped and needed to be reconnected, and it looks like there were some settings I could/should change for nginx, so I did, and we'll see if it helped...

But in the process, I forgot that the demo branch in this repo had a bunch of Celery and Twitter stuff, that hadn't been added to vine-secrets, so when I re-deployed, everything broke, and the site was down for a few minutes, and ugh I should be more careful.

These changes should fix it so that the site can now run on demo again. Note that this includes a totally useless Twitter login button on all pages, but whatever. The Celery stuff should be ready, but I'm not going to try to add it to Supervisor now.