Update score_edges.py for new leaf_component
The leaves now handle adding and removing the edges from the database, so we need to remove that from here.

Update the messaging to be more consistent and match the changes to the leaf messaging.

Since this now no longer changes all edges, we need to use a different method of cycling through all the old edges one at a time.

Note that I'm nore sure why we weren't deleting scores from the dict if the check_score failed in update_next_old_edge. This would otherwise get handled at the end of update_next_new_edge, but I'm pretty sure we should be removing scores here that are too low.