Add JWChat web-based demos to vine_web cookbook
This uses a new recipe called automatically by default to prepare all the necessary static files for the JWChat web-based demo. It's a pretty straightforward recipe that gets the code, installs the necessary Perl modules, uses execute to run "make", and uses a template for the config file.

Note that it seems like Perl comes pre-installed on the distro of Ubuntu that I'm using? I'll worry about it if it breaks.

Also note that we're now using a Chef template to dynamically render the domain-specific variables in JWChat's config.js file, rather than relying on that hacky bit in the nginx.conf to branch between two files based on the domain, and that hacky perl regex.

Also updated the README.md to no longer have the JWChat/Perl instructions, yay!

Also changed the xmlrpc_server constant in the databag to xmpp_server, since it's used for other things. See https://github.com/lehrblogger/vine-chef/commit/183bc460399c410f2efdc0a3ab61befbf2838a70