Properly fetch_follows & score_edges for new users
Before, it was only scoring edges for new users, which meant that it wasn't fetching their Twitter follows and using that to pre-populate their social graph until they logged in again or until the daily task ran, which was a pretty bad experience. Now, both tasks fire with both types of new users (those who had signed in before, and those who hadn't (I think)), and should be complete by the time the user signs in with a client.