Don't launch JWChat with invalid passwords, WEIRD
This was really weird, since it made the whole page weirdly impossible to interact with if the passwords don't match or were blank. Turned out there were a couple of issues. First, it wasn't checking to make sure that the passwords were valid before trying to launch the web client, which of course will fail if the new password is different than the old password, because ejabberd will have rejected the change.

Secondly and more subtly, in a new incognito Chrome window (with no cache), it tried to launch JWChat, but had to load an html page from the server, which it did fine, but before it finished, the page submitted the form and returned the new page, which cleared the global variables that the JWChat page was expecting. To fix this, we're waiting for that page to load before actually submitting the form, which is pretty clean/straightforward despite the surrounding global-variable onclick ugliness.

One superweird thing is that the Chrome window was freezing from the missing JID alert thrown by jwchat.html, and I'd have to open a new tab with the button, or close the window entirely. Ugh.