Add support for invites tied to Twitter users
I want users to be able to tweet invite links @ specific Twitter users they want to invite, so that I can next add links directly to specific conversations. But to make that work, I need to be able to restrict a certain invite to only work with a specific Twitter username. This updates the various flows accordingly.