Basic /index support for new invite/invitee tables
Oh wow am I not happy with SQLAlchemy right now. It took FOREVER to figure out how to do the first part. Initially I wanted to use a subquery to count, but that syntax was impossible to find/understand, and I couldn't figure out how to get it to use a variable from the main query in the subquery on my own. Thankfully Adam has used SQLAlchemy and suggested I use a join instead (ugh I wish I actually knew SQL) and helped me get it working.
And then the second part should have been easy, but for some reason I think the SQLAlchemy session in my python shell was somehow out of date, and didn't have all of the rows in the DB that it was supposed to, and was thus not finding the ones I was expecting. FML.

Things are in a broken state with this commit, but whatever.