Finish basic support for using multi-use invites
This required rewriting a bunch of queries, but the functionality should basically be the same. I'll add UI for the multi-use invites later, but the admin /commands on the leaves should suffice for now.

Also, the edge calculation script had to be tweaked slightly to account for the new tables.