Add screencast.gif to logged-out pages, fix CSS
As a result, I was able to remove the header banner entirely, and move the sign up/sign in stuff into the right column, which is great. The chat bubble animation is also gone (HTML/CSS/JS).

I also restyled a ton of stuff, and deleted unused CSS, and cleaned up the templates a bit.