Add ToS/PP confirmation to homepage, other cleanup
I was originally looking for a bug where the same invite code could be used twice, but I couldn't find it, and cleaned up some stuff in the process.

This undoes much of 22ef02c87e474729addda04c548e182275be5860 and simplifies the signup/login flow a bit, which is good. In the old header banner there was no space for the legal copy, so I added a session variable called 'terms_accepted' to indicate whether a user had selected the terms on the signin page. This was a bit messy, and I'll need that copy on the home page anyway, so I might as well add it for now.

There are also scattered CSS and template changes, but nothing major.