Rewrite email list form for MailChimp, not Google
Now, the form posts a request to my server, which in turn uses the MailChimp API to add the user as a subscriber on my list. This means I no longer need the Google Spreadsheet that I was using to keep track of the emails before, which saves me a step whenever I want to send an email out.

For the sake of simplicity/avoiding doing validation on both the client and server, I'm now using a Flask form instead of the JS thing Joseph wrote, so I rewrote the CSS and got rid of the JS.

Note there are some tiny Chef data bag and recipe changes for this too, since I don't want to hard-code my MailChimp API key.

Later, I'll add existing users to a MailChimp list as well using the API.