Use jQueryUI tabs for /help, support URL fragments
There was a jQueryUI file before, but it seemed not to be working, so now this is using a new one.

I'm also using this technique for the push state history thing:
	http://lea.verou.me/2011/05/change-url-hash-without-page-jump/
And this one for preventing it from headbutting the browser and scrolling past the header.
	http://nicolasgallagher.com/jump-links-and-viewport-positioning/demo/

Ugh CSS, ugh JS.