Stop /help URL fragments from scrolling to anchors
The problem is that if you go to pages like https://dev.vine.im/help#messages, it would scroll past the header to the beginning of the <ol>, which is no good from a UX perspective. I had tried a bunch of CSS things to fix it, and I *thought* the one that I am removing in this commit worked, but it didn't, because it covered the header and made it unclickable. Ugh.

So, I tried a bunch more CSS things, and a bunch of JS things (note you can't prevent URL fragment behavior with JS, apparently: http://stackoverflow.com/questions/3503559/prevent-default-hash-behavior-on-page-load#comment3685156_3503621), and Googled all over, but found nothing, and finally came up with what is here.

I think this is what's going on/why this works:
	1) document.ready starts, and creates the tabs. I THINK that when it creates the tabs for the first time, it looks at the URL fragment, and figures out which tab index, if any, should be specified as active.
	2) document.ready then renames all of the hrefs and ids to have "-setup" at the end, which prevents the URL fragment from matching them.
	3) The browser now(??) decides what the scroll position should be, based on the URL fragment. Since the ids all have "-setup" appended to them, nothing happens.
	4) window.load fires, and changes all of hrefs and ids back.
	5) window.load refreshes the tabs, so everything is as it should be and URLs can be clicked/refreshed/etc normally.

This may not work in Firefox, but whatever.