Use templates, not jQuery-UI tabs, for /help page
Loading all of the images for all of the clients was sometimes slow, and resulted in sporadically-buggy tab behavior. Rather than figure out some JS lazy image loading, it seemed easier to put each set of instructions on its own separate page, and to handle the routing in Flask. This makes the templates a tiny bit more complicated, but simplifies the JS and CSS, so I'm happy.