Add /signin page handler, improve 404 & 500 pages
On Jorge's suggestion, I added the /signin handler so that when users try to go to pages that they need to be logged-in to view, they realize it, rather than just getting redirected to the index page.

Also cleaned up and re-themed the 404 and 500 error pages, and put the wonderland easter egg in a snippet to make it easier to include.