Improve :visited CSS on /about Tumblr button, ARGH
http://mlkshk.com/r/PPML

So, after an extraordinary amount of pain, I found https://developer.mozilla.org/en-US/docs/Web/CSS/:visited:

	Note: For privacy reasons, browsers strictly limit the styles you can apply using an element selected by this pseudo-class: only color, background-color, border-color, border-bottom-color, border-left-color, border-right-color, border-top-color, outline-color, column-rule-color, fill and stroke.

I think this means it's impossible to have a button that looks like a Bootstrap button normally, and not like a Bootstrap button when :visited, because there's no way to change the background or background-image properties with the :visited pseudo selector.

I tried TONS of stuff. It seems like applying the background style to the :link pseudoselector should work, but it instead applied it in both that and the :visited state.

I guess what I actually want is impossible without Javascript of some sort, but I guess this is a good enough compromise for now.