Fix deploy_wrapper version issue again (see 6f0cb0eda28afaef396d0be4dd9bc9414e428ffe)
I don't really understand why this is happening, but the error message looks like this:

	[2013-05-20T07:19:53+00:00] INFO: HTTP Request Returned 412 Precondition Failed: {"unsatisfiable_run_list_item":"(vine_web >= 0.0.0)","non_existent_cookbooks":[],"most_constrained_cookbooks":["Package deploy_wrapper"],"message":"Unable to satisfy constraints on cookbook deploy_wrapper, which does not exist, due to run list item (vine_web >= 0.0.0). Run list items that may result in a constraint on deploy_wrapper: [(vine_web = 1.0.1) -> (deploy_wrapper = 0.0.4)]"}

I tried to fix it by specifying versin >= 0.0.2, but that didn't work. Then I went to the communtiy site and downloaded 0.0.2, rather than 0.0.4 (not sure why this happened in the first place), and put that on my Hosted Chef, and it seems to have worked. Weird, whatever.