Fix CPU max on login w/ score_edges(user_id=None)
I think, at least. When a user logged in to the site, the first time the score_edges task got called, it was getting a user_id of None, which means it will do it for all users, which is kind of CPU intensive. If this happens twice in reasonably quick succession, it can max the MySQL CPU, which was sending alerts.

I think this was happening because the argument passing in the celery chain was not working, so the user_id from fetch_follows wasn't making it into score_edges for the first time it ran, but it was getting passed explicitly fine for subsequent times.

I'm also seeing many score_edges tasks for each login, and I'm not sure why this is. To make things a little more gentle, I got rid of the last call at 6 hours.

Also, note that I changed the SQS retention period from 4 days to 10 minutes, in case things were getting stuck and done twice, or something. I'm not using SQS for anything that important, so I don't need old stuff sitting there.