Fix jwchat issue where perl couldn't find Regex lib
Argh this was annoying. It seemed like the cpan cookbook was installing the libraries to a place that perl wasn't looking. I tried *many* things, but for some reason I couldn't install_base wasn't having any effect. I think this is only an issue on dev, since prod installs to the right place? We'll see...